+++ src/main/java/com/platinumg17/rigoranthusemortisreborn/magica/common/potions/SummoningCooldownEffect.java
@@ -10,7 +10,7 @@ import java.util.List;
 
 public class SummoningCooldownEffect extends Effect {
     protected SummoningCooldownEffect() {
-        super(EffectType.HARMFUL, 2039587);
+        super(EffectType.NEUTRAL, 2039587);
         setRegistryName(EmortisConstants.MOD_ID, "summoned_servant");
     }